requirements.txt
selenium==4.5.0
undetected_chromedriver==3.4.7
schedule==1.2.0
Pillow==10.1.0
onnxruntime==1.18.1
numpy==1.26.2
python-dotenv==1.0.0

